Resolve the Schedule class from the container instead of DI, fixes #30
The Schedule class is resolved from the Container with the global app() function when needed in the
ScheduleList::all method instead of Dependecy injection in the construction.
This is to make sure that Laravel has resolved the class itself and filled in with all schduled tasks
and we get the same instance.